Change: Support objects with more than 65,535 vertices in a single mesh
https://bveworldwide.forumotion.com/t2350-some-x-file-s-polygons-are-broken#21543

+++ source/OpenBveApi/Objects/Helpers/MeshBuilder.cs
@@ -64,7 +64,7 @@ namespace OpenBveApi.Objects
 					Object.Mesh.Faces[mf + i] = Faces[i];
 					for (int j = 0; j < Object.Mesh.Faces[mf + i].Vertices.Length; j++)
 					{
-						Object.Mesh.Faces[mf + i].Vertices[j].Index += (ushort) mv;
+						Object.Mesh.Faces[mf + i].Vertices[j].Index += mv;
 					}
 
 					Object.Mesh.Faces[mf + i].Material += (ushort) mm;
